refactoring complete. Tags and Demographics actually removed and form inputs moved to Report component. context connected and form submission successfully updates state
